Initial implementation of heavy truck, with telemetry-based rev-matching for ATS/ETS2. See #12.

+++ BonusFFB/DeviceInfo.h
@@ -82,4 +82,4 @@ DeviceInfo* getDeviceFromGuid(QList<DeviceInfo>*, QUuid);
 BOOL CALLBACK enumDevicesCallback(const DIDEVICEINSTANCE*, VOID*) noexcept;
 BOOL CALLBACK enumAxesCallback(const DIDEVICEOBJECTINSTANCE*, VOID*) noexcept;
 
-double scaleRangeValue(long, long, long);
+double scaleRangeValue(double, double, double);
